chore: add Rust build artifacts for dgt_harvest_rust release compilation with image processing dependencies (release/.fingerprint with adler2-7615ae9ed3220a7e/autocfg-e137e6ca9ebf7b6a/bit_field-96a9a9d7bf0f109a/bitflags-4c9efaf164c422a3/bytemuck-54d982dc95b53e6b/byteorder-f81eec70149406a7/cfg-if-3cfe40de624c43d8/color_quant-b9f61f05c9100261/crc32fast-77ce778ce1e215c0/crossbeam-deque-623c75a845271d46/crossbeam-epoch-4b0c6ac9abeea1a4/crossbeam-utils-5e82e98ecbd4dd4b dependency metadata with invoked.timestamp/

+++ rust_sprite_scanner.py
@@ -64,7 +64,7 @@ class RustSpriteScanner:
             # Define Rust function signature
             scan_function = rust_lib.scan_sprite_for_chest
             scan_function.argtypes = [ctypes.c_char_p, ctypes.c_uint, ctypes.c_uint]
-            scan_function.restype = ctypes.c_float
+            scan_function.restype = ctypes.c_double
             
             # Convert pixels to bytes
             pixels_ptr = ctypes.create_string_buffer(pixels)
